Initial Commit
Sorry, should have put this into a repo earlier really.

The quicktalk plugin tries to make it easier to do lipsynch
in blender.

Here we have the basic panel with buttons, a function to
create a control panel in an armature for an object with
shape-keys to control those shape-keys and also a class
to read and parse a script file to determine who says
what.

Next up:
  * The 'guess dialogue markers' button needs to add some markers