Add the actual f-curve plots to say the words.
And my first example/test code seems to work a
treat at that and not even need any squashing
or stretching to get the words to match properly.

Win!

Also a fix to ensure we only build panel-bones
out of the shape-keys we really *want* to, so
any shape-keys whose names starts with an "!"
we ignore.

Also found that on one of my test models the
armature has a scale, 0.1 in fact, which meant
the panel wasn't being built in the right
place.

Fixed that.