Allow the phoneme "TH" to be used as a separate shape.
Blender Artist's user "TCooper" suggested that,
especially for characters with a lisp, it was
important for the "TH" phoneme to have it's own
shape-key.

So here we expand the list of shape-keys/bones
to include that. If you don't add it, things
will work as before and the "ETC" shape will
be used.