System to allow us to use MHX2 imported models.
Which presumably come in from Makehuman in general.

We just turn on "record-keys", move to
the frame and call the mhx2 "SetViseme"
function basically.

Annoying that the dictionary is so deeply
embedded into the load, for it would be
better if we could fit the proper visemes
in from the phonemes rather than relying
on translation via my standard set which
we still do at the moment.

Hummm.